backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pinecone-client==3.0.0
openai==1.10.0
python-dotenv==1.0.0
opencv-python>=4.10.0
scenedetect[opencv]==0.6.3
pypdf==4.0.0
python-docx==1.1.0
python-magic==0.4.27
pillow==10.2.0
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3
